Version Compatibility (#23)
* Transition @kksh/api to use TypeScript completely

without bundling to js at all. There was some problem with bundling with rollup when I use json plugin.

* feat: add auto version patch to API package

instead of exporting directly from package.json, now use a script to add version to typescript file during build

* test: make unit test quite

Avoid too much log in CI

* ci: implement version checking

* chore: add @types/semver to api pkg

+++ packages/api/build.ts
@@ -1,7 +1,13 @@
+import fs from "fs"
 import { $ } from "bun"
 
-await $`rm -rf dist`
-await $`pnpm build:rollup`
+// add package version
+
+if (fs.existsSync("dist")) {
+	await $`rm -rf dist`
+}
+fs.mkdirSync("dist")
+// await $`pnpm build:rollup`
 // await $`cp ../schema/manifest-json-schema.json ./dist/schema.json`
 await $`bun ../schema/scripts/print-schema.ts > dist/schema.json`
 
@@ -10,3 +16,5 @@ const schemaFile = Bun.file("dist/schema.json")
 if (!schemaFile.exists()) {
 	throw new Error("schema.json not found")
 }
+
+await $`bun patch-version.ts`
